Adapting sharpen to new matrix returned from convolution

+++ sharpen.m
@@ -3,5 +3,5 @@ function [x, mask] = sharpen(img)
        -1,  8, -1;
        -1, -1, -1];
   mask = convolution(img, w);
-  x = img + mask;
+  x = img + uint8(mask);
 endfunction
